New client structure has been made short
Previously two separate but inter-linked entities were designed for client and the heartbeat, but now, both the entities has been merged to a single entity. A new boolean field called 'alive' has been introduced where the heartbeat result will be stored. And as the result is updated, the 'updated_at' field is stored with the timestamp of that time so that we can trace if the server is alive.